requirements.txt
langchain
langchain-community
langchain_groq
langchain-chroma
chromadb
streamlit
pandas
python-dotenv
sentence-transformers
langchain_huggingface
